+++ code/espurna/terminal.cpp
@@ -19,7 +19,10 @@ Copyright (C) 2020 by Maxim Prokhorov <prokhorov dot max at outlook dot com>
 #include "utils.h"
 #include "wifi.h"
 #include "ws.h"
+
 #include "libs/URL.h"
+#include "libs/StreamAdapter.h"
+#include "web_asyncwebprint_impl.h"
 
 #include <algorithm>
 #include <vector>
@@ -482,58 +485,6 @@ void terminalOK() {
 void terminalError(const String& error) {
     terminalError(_io, error);
 }
-
-template <typename T>
-struct StreamAdapter : public Stream {
-    StreamAdapter(Print& writer, T&& begin, T&& end) :
-        _writer(writer),
-        _current(std::forward<T>(begin)),
-        _begin(std::forward<T>(begin)),
-        _end(std::forward<T>(end))
-    {}
-
-    int available() final override {
-        return (_end - _current);
-    }
-
-    int peek() final override {
-        if (available() && (_end != (1 + _current))) {
-            return *(1 + _current);
-        }
-        return -1;
-    }
-
-    int read() final override {
-        if (_end != _current) {
-            return *(_current++);
-        }
-        return -1;
-    }
-
-    void flush() final override {
-// 2.3.0  - Stream::flush()
-// latest - Print::flush()
-#if not defined(ARDUINO_ESP8266_RELEASE_2_3_0)
-        _writer.flush();
-#endif
-    }
-
-    size_t write(const uint8_t* buffer, size_t size) final override {
-        return _writer.write(buffer, size);
-    }
-
-    size_t write(uint8_t ch) final override {
-        return _writer.write(ch);
-    }
-
-    protected:
-
-    Print& _writer;
-
-    T _current;
-    T const _begin;
-    T const _end;
-};
 
 void terminalSetup() {
